Move the DateTimePicker and InputDropdown to common_widgets
